Resuming from saved models
This should allow evolution to resume from a previously saved point. It seems to load the models and start evaluating them, but I haven't completed a full generation to see what happens. I believe this process will re-evaluate and re-save the models, overwriting the previous ones. The parents should be the same, but the children are not loaded, and actually get overwritten.

Also, I'm not currently loading the saved behavior archive,  which is a no-go #38