Updated requirements
It seems to be very important to have gym-retro version 0.7.0

I removed the tensorflow version restriction because it was causing problems, but I don't think we really use it anyway since the move to PyTorch. Might remove entirely.

baselines is needed, though I'm not sure getting the default version is a great idea.

Also added comment about getting and installing the retro-contest repo, since this is definitely needed ... can this simply be done with pip?