Fixed error requiring tensor cloning
This code definitely worked before on other systems. I'm not sure why it is causing an error now. Maybe a change in the CUDA version, the Pytorch version, or being used on a different machine, but cloning the tensor to a temporary variable fixes the problem (the error message was very helpful).